3d plot, waiting on accelero data.
tested the thing with different accelerometer data, rest can come in serially or something and we can just update based on serially reading in the data. can make a functino to read in serial data or just pull it from somewhere and pass in the values to the update function, and set ys1,ys2,ys3 to it.